iommufd/viommu: Introduce IOMMUFD_OBJ_VQUEUE and its related struct
Add a new IOMMUFD_OBJ_VQUEUE with an iommufd_vqueue structure to represent
a queue type of physical HW passed to or shared with a user space VM. This
queue, now a vQUEUE object, is a subset of vIOMMU virtualization resources
of a physical IOMMU's, such as:
 - NVIDIA's virtual command queue
 - AMD vIOMMU's command buffer

Inroduce a struct iommufd_vqueue and its allocator iommufd_vqueue_alloc().
Also, add a pair of viommu ops for iommufd to forward user space ioctls to
IOMMU drivers.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>