iommufd/viommu: Introduce IOMMUFD_OBJ_VQUEUE and its related struct
Add a new IOMMUFD_OBJ_VQUEUE with an iommufd_vqueue structure representing
a HW-accelerated queue type of physical IOMMU's passed to a user space VM.
This vQUEUE object, is a subset of vIOMMU resources of a physical IOMMU's,
such as:
 - NVIDIA's Virtual Command Queue
 - AMD vIOMMU's Command Buffer, Event Log Buffer, and PPR Log Buffer

Introduce a struct iommufd_vqueue and an allocator iommufd_vqueue_alloc().
Also, add a pair of viommu ops for iommufd to forward user space ioctls to
IOMMU drivers.

Reviewed-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: Pranjal Shrivastava <praan@google.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>