cover-letter: iommu: Support IOMMU_RESV_SW_MSI with nesting
The dma-iommu needs to find the correct domain for MSI mapping. With an
IOMMU_DOMAIN_NESTED, the mapping resides in its parent paging domain.

Add a get_msi_mapping_domain op for drivers to return paging domains.

Add an iommufd selftest coverage for that, by doing a loopback test.

Add arm_smmu_get_msi_mapping_domain in the SMMUv3 driver so its nesting
feature could work with MSI correctly.

This is based on top of the reserved-IOVA change:
https://lore.kernel.org/all/20240802053458.2754673-1-nicolinc@nvidia.com/
And Jason's SMMUv3 nesting series:
https://lore.kernel.org/all/0-v1-54e734311a7f+14f72-smmuv3_nesting_jgg@nvidia.com/

This series is on Github:
https://github.com/nicolinc/iommufd/commits/iommufd_nesting_sw_msi/

[changelog]
v3:
 * Refined PATCH-2 commit message
 * Added domain->ops check in PATCH-2
 * Added PATCH-4 to implement in SMMUv3 driver
v2:
 https://lore.kernel.org/all/cover.1722644866.git.nicolinc@nvidia.com/
 * Resent with a proper bug fix.

Thanks
Nicolin