iommufd/selftest: Add refcount to mock_iommu_device
The viommu->iommu_dev pointer has no guarantee of its life cycle after it
is copied from the idev->dev->iommu->iommu_dev. The end of the life cycle
means the iommu device is unplugged from the system. This rarely happens,
yet selftest does unplug the iommu device during the exit routine.

Track the user count of the iommu_dev. Delay the exit routine if refcount
is unbalanced. The refcount inc/dec will be added in the following patch.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>