iommufd/fault: Add an iommufd_fault_init() helper
The infrastructure of a fault object will be shared with a new vIRQ object
in a following change. Add a helper for a vIRQ allocator to call it too.

Reorder the iommufd_ctx_get and refcount_inc to keep them symmetrical with
the iommufd_fault_fops_release().

Since the new vIRQ object doesn't need "response", leave the xa_init_flags
in its original location.

Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>